backend/requirements.txt
fastapi[standard]
pandas
python-dotenv